requirements.txt
flask
supabase
python-dotenv
qrcode
pillow
gunicorn
sendgrid